Move BcrFilter to its own file
