+++ Unit4/Tests/BcrFilterTests.cs
@@ -67,21 +67,4 @@ namespace Unit4.Automation.Tests
             Assert.That(filter.Use(bcr).Lines, Has.Count.EqualTo(1));
         }
     }
-
-    internal class BcrFilter : IBcrMiddleware
-    {
-        private readonly BcrOptions _options;
-
-        public BcrFilter(BcrOptions options)
-        {
-            _options = options;
-        }
-
-        public Bcr Use(Bcr bcr)
-        {
-            var newBcr = new Bcr();
-            newBcr.Lines = _options.Tier2 == null ? bcr.Lines.ToList() : bcr.Lines.Where(x => x.CostCentre.Tier2.Equals(_options.Tier2)).ToList();
-            return newBcr;
-        }
-    }
 }
