+++ Unit4/Unit4.csproj
@@ -103,6 +103,7 @@
     <Compile Include="NullRunner.cs" />
     <Compile Include="BcrReportRunner.cs" />
     <Compile Include="BcrReader.cs" />
+    <Compile Include="BcrFilter.cs" />
     <Compile Include="ReportRunnerFactory.cs" />
     <Compile Include="BcrReport.cs" />
     <Compile Include="Cache.cs" />
